# backend/vc-management-service/management-service-swagger.yml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: VC Management
  description: verifiable credentials management service
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/context:
    post:
      tags: upload a context
      summary: Upload a context json to minio
      consumes:
        - multipart/form-data
      produces:
        - application/json
      operationId: addContext
      parameters:
        - in: formData
          name: files
          type: file
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UploadTemplateSuccess'
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  /v1/context/:osid:
    get:
      tags:
        - get context
      summary: Get context in json ld format
      consumes:
        - application/json
      produces:
        - application/ld + json
      operationId: getContext
      parameters:
        - in: path
          name: osid
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '400':
          description: Invalid input
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  UploadTemplateSuccess:
    type: object
    properties:
      uploadTemplateResponse: 
        type: object
        properties:
          documentLocations: 
            type: array
            items:
              type: string
          errors: 
            type: array
            items:
              type: string
  Error:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
host: divoc.xiv.in
schemes:
  - https
basePath: /vc-management/